# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LSP UGM API Documentation",
        "description": "API Documentation for LSP UGM Application",
        "contact": {
            "name": "Administrator",
            "email": "admin@lsp-ugm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "LSP UGM API Server"
        }
    ],
    "paths": {
...
        "/asesmen/mapa02/{id_asesi}": {
            "get": {
                "tags": [
                    "MAPA.02"
                ],
                "summary": "Mendapatkan data formulir MAPA.02 untuk asesi",
                "description": "Get Mapa02 data for an Asesi",
                "operationId": "cc45bd19c04eb33149fdadfc6f0a811f",
                "parameters": [
                    {
                        "name": "id_asesi",
                        "in": "path",
                        "description": "ID asesi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data formulir MAPA.02",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "general_info": {
                                                    "properties": {
                                                        "nama_asesi": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "nama_asesor": {
                                                            "type": "string",
                                                            "example": "Jane Smith"
                                                        },
                                                        "judul_skema": {
                                                            "type": "string",
                                                            "example": "Programmer"
                                                        },
                                                        "kode_skema": {
                                                            "type": "string",
                                                            "example": "SKM-001"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "mapa02": {
                                                    "properties": {
                                                        "muk_ceklis_observasi": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "muk_tugas_praktik_demonstrasi": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "muk_pertanyaan_tertulis_esai": {
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "muk_pertanyaan_lisan": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "muk_ceklis_verifikasi_portfolio": {
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "muk_ceklis_meninjau_materi_uji": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "waktu_tanda_tangan_asesor": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2025-05-15"
                                                        },
                                                        "tanda_tangan_asesor": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "signature/ttd_asesor.png"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "record_exists": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asesi tidak ditemukan atau belum memiliki rincian asesmen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak ditemukan"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Formulir prasyarat belum diselesaikan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Formulir MAPA-01 (Merencanakan Aktivitas) belum diselesaikan"
                                        },
                                        "missing_form": {
                                            "type": "string",
                                            "example": "mapa01"
                                        },
                                        "display_name": {
                                            "type": "string",
                                            "example": "MAPA-01 (Merencanakan Aktivitas)"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/asesmen/mapa02/save": {
            "post": {
                "tags": [
                    "MAPA.02"
                ],
                "summary": "Menyimpan data formulir MAPA.02",
                "description": "Create or update Mapa02 data",
                "operationId": "6e260efdd6c4759f9d889ec7f0b76f9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_asesi",
                                    "id_asesor",
                                    "muk_ceklis_observasi",
                                    "muk_tugas_praktik_demonstrasi",
                                    "muk_pertanyaan_tertulis_esai",
                                    "muk_pertanyaan_lisan",
                                    "muk_ceklis_verifikasi_portfolio",
                                    "muk_ceklis_meninjau_materi_uji"
                                ],
                                "properties": {
                                    "id_asesi": {
                                        "type": "string",
                                        "example": "ASESI2025XXXXX"
                                    },
                                    "id_asesor": {
                                        "type": "string",
                                        "example": "ASESOR2025XXXXX"
                                    },
                                    "muk_ceklis_observasi": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "muk_tugas_praktik_demonstrasi": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "muk_pertanyaan_tertulis_esai": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "muk_pertanyaan_lisan": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "muk_ceklis_verifikasi_portfolio": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "muk_ceklis_meninjau_materi_uji": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "is_signing": {
                                        "description": "Set to true if the asesor is signing the form",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data MAPA.02 berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data MAPA.02 berhasil disimpan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asesi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak ditemukan"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Asesor tidak ditugaskan untuk asesi ini",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesor tidak ditugaskan untuk asesi ini"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 403
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal atau formulir prasyarat belum diselesaikan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            }
        }
    },
...
}